refactor: Recursive copy exported to a filesystem utils class. Unit tests added for it.
