+++ typescript/polyglotpkg/.gitignore
@@ -8,3 +8,4 @@ node_modules/
 **/.DS_Store
 dist/
 **/polyglot-cache/
+test/unit/file-system/copy-files/destination
