+++ typescript/polyglotpkg/src/strategies/powershell.ts
@@ -17,9 +17,10 @@ import globby from 'globby';
 
 import { Logger } from "winston";
 import { BaseStrategy } from "./base";
-import { cpSync, run } from "../lib/utils";
+import { run } from "../lib/utils";
 import { ActionOptions, PlatformDefinition, Events } from "../lib/model";
 import { mkdirSync, readFileSync } from 'fs';
+import { cpSync } from '../lib/file-system';
 
 export class PowershellStrategy extends BaseStrategy {
 
@@ -79,7 +80,6 @@ export class PowershellStrategy extends BaseStrategy {
 
 	private async compile(source: string, destination: string) {
 		this.logger.info(`Compiling project...`);
-        // TODO: Replace with cpSync from the Node.js "fs" API when TypeScript and @types/node are bumped
         cpSync(source, destination);
 		this.logger.info(`Compilation complete`);
 	}
