+++ typescript/polyglotpkg/test/unit/file-system/copy-files/source/file1
@@ -1,0 +1,1 @@
+file 1 contents
