+++ typescript/polyglotpkg/test/unit/file-system/copy-files/source/file2
@@ -1,0 +1,1 @@
+file 2 contents
